requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# File processing
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.1.0
python-magic==0.4.27

# DeepSeek/OpenAI API
openai==1.12.0

# HTTP client
httpx==0.25.1
aiohttp==3.9.1

# Utilities
pydantic[email]==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
email-validator>=2.0.0
dnspython>=2.0.0
celery==5.3.4

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
mypy==1.7.0